+++ infrastructures/symfony/Messenger/Executor.php
@@ -1,0 +1,60 @@
+<?php
+
+/*
+ * East Foundation.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the MIT license and the version 3 of the GPL3
+ * license that are bundled with this package in the folder licences
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to richarddeloge@gmail.com so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) 2009-2021 EIRL Richard Déloge (richarddeloge@gmail.com)
+ * @copyright   Copyright (c) 2020-2021 SASU Teknoo Software (https://teknoo.software)
+ *
+ * @link        http://teknoo.software/east Project website
+ *
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richarddeloge@gmail.com>
+ */
+
+declare(strict_types=1);
+
+namespace Teknoo\East\FoundationBundle\Messenger;
+
+use Teknoo\East\Foundation\Manager\ManagerInterface;
+use Teknoo\Recipe\BaseRecipeInterface;
+
+/**
+ * @copyright   Copyright (c) 2009-2021 EIRL Richard Déloge (richarddeloge@gmail.com)
+ * @copyright   Copyright (c) 2020-2021 SASU Teknoo Software (https://teknoo.software)
+ *
+ * @link        http://teknoo.software/east Project website
+ *
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richarddeloge@gmail.com>
+ */
+class Executor
+{
+    private ManagerInterface $manager;
+
+    public function __construct(ManagerInterface $manager)
+    {
+        $this->manager = $manager;
+    }
+
+    /**
+     * @param array<string, mixed> $workPlan
+     */
+    public function execute(BaseRecipeInterface $recipe, array $workPlan): self
+    {
+        $this->manager->read($recipe);
+
+        $this->manager->process($workPlan);
+
+        return $this;
+    }
+}